refactor: cDrawDown to own file
