+++ Model/cDrawDown.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace VisualHFT.Model
+{
+    public class cDrawDown
+    {
+        public DateTime Date { get; set; }
+        public decimal DrawDownAmmount { get; set; }
+        public decimal DrawDownPerc { get; set; }
+        public int DrawDownHours { get; set; }
+    }
+}
