+++ Model/cEquity.cs
@@ -21,11 +21,4 @@ namespace VisualHFT.Model
                 return null;
         }
     }
-    public class cDrawDown
-    {
-        public DateTime Date { get; set; }
-        public decimal DrawDownAmmount { get; set; }
-        public decimal DrawDownPerc { get; set; }
-        public int DrawDownHours { get; set; }
-    }
 }
